ref: extract function to fetch softwares
