+++ client/src/pages/[locale]/softwares/[id].astro
@@ -2,20 +2,15 @@
 import Layout from "@/layouts/Layout.astro";
 import SoftwareDetail from "@/components/SoftwareDetail.astro";
 import { getRelativeLocaleUrl } from "astro:i18n";
-import  { type Lang, useTranslations} from "@/i18n/utils";
-import type { PubliccodeYml, Software } from "@/types/software";
+import { type Lang, useTranslations } from "@/i18n/utils";
+import type { PubliccodeYml, Software } from "@/lib/software";
 import yaml from "js-yaml";
 import { hashUrl } from "@/utils/hash";
-
-const apiBaseUrl = import.meta.env.API_BASEURL || "http://localhost:3000/v1";
+import { getSoftwares } from "@/lib/software";
 
 export async function getStaticPaths() {
-  const apiBaseUrl = import.meta.env.API_BASEURL || "http://localhost:3000/v1";
-  const locales = ['en', 'de', 'fr', 'it'];
-
-  const response = await fetch(`${apiBaseUrl}/software?page[size]=10000`);
-  const data = await response.json();
-  const softwares: Software[] = data.data;
+  const softwares = await getSoftwares();
+  const locales = ["en", "de", "fr", "it"];
 
   return softwares.flatMap((software) =>
     locales.map((locale) => ({
@@ -26,15 +21,15 @@ export async function getStaticPaths() {
       props: {
         software,
       },
-    }))
+    })),
   );
 }
 
-const rawSoftware = Astro.props.software;
-
 const software: Software = {
-  ...rawSoftware,
-  publiccode: yaml.load(rawSoftware.publiccodeYml as unknown as string) as PubliccodeYml,
+  ...Astro.props.software,
+  publiccode: yaml.load(
+    Astro.props.software.publiccodeYml as unknown as string,
+  ) as PubliccodeYml,
 };
 
 const lang = Astro.currentLocale as Lang;
@@ -59,11 +54,11 @@ const homeUrl = getRelativeLocaleUrl(lang);
           xmlns="http://www.w3.org/2000/svg"
           d="m8.306 5.116-4.066 7.044 4.066 7.044.65-.375-3.633-6.294h15.187v-.75h-15.187l3.633-6.294z"
         ></path></svg
-      ><span class="btn__text">{t('software.back')}</span></a
+      ><span class="btn__text">{t("software.back")}</span></a
     >
   </div>
   <div class="container vertical-spacing">
-    <h1 class="h2">{t('software.title')}</h1>
+    <h1 class="h2">{t("software.title")}</h1>
     <SoftwareDetail software={software} />
   </div>
 </Layout>
